Separation of  VisitService form ClinicService
