+++ src/main/java/org/springframework/samples/petclinic/service/VisitService.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2002-2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.samples.petclinic.service;
+
+import java.util.Collection;
+
+import org.springframework.dao.DataAccessException;
+import org.springframework.samples.petclinic.model.Visit;
+
+
+/**
+ * Mostly used as a facade for the visit controller
+ *
+ * @author Nicolas Depelteau
+ */
+public interface VisitService {
+
+	Collection<Visit> findVisitsByPetId(int petId);
+	Visit findVisitById(int visitId) throws DataAccessException;
+	Collection<Visit> findAllVisits() throws DataAccessException;
+	void saveVisit(Visit visit) throws DataAccessException;
+	void deleteVisit(Visit visit) throws DataAccessException;
+}
